+++ demo/generic_data_structures/linked_list.h
@@ -7,8 +7,8 @@
 typedef struct list ioopm_list_t;
 typedef struct iter ioopm_list_iterator_t;
 typedef struct list_option ioopm_list_option_t;
-typedef bool ioopm_int_predicate(int index, elem_t value, void *extra);
-typedef void ioopm_apply_int_function(int index, elem_t *value, void *extra);
+typedef bool ioopm_int_predicate(size_t index, elem_t value, void *extra);
+typedef void ioopm_apply_int_function(size_t index, elem_t *value, void *extra);
 
 struct list_option
 {
@@ -43,7 +43,7 @@ void ioopm_linked_list_prepend(ioopm_list_t *list, elem_t value);
 /// @param list the linked list that will be extended
 /// @param index the position in the list
 /// @param value the value to be inserted
-void ioopm_linked_list_insert(ioopm_list_t *list, int index, elem_t value);
+void ioopm_linked_list_insert(ioopm_list_t *list, size_t index, elem_t value);
 
 /// @brief Remove an element from a linked list in O(n) time.
 /// The valid values of index are [0,n-1] for a list of n elements,
@@ -52,7 +52,7 @@ void ioopm_linked_list_insert(ioopm_list_t *list, int index, elem_t value);
 /// @param index the position in the list
 /// @param success a stack variable pointer to indicate success or failure
 /// @return the value removed
-elem_t ioopm_linked_list_remove(ioopm_list_t *list, int index, bool *success);
+elem_t ioopm_linked_list_remove(ioopm_list_t *list, size_t index, bool *success);
 
 /// @brief Retrieve an element from a linked list in O(n) time.
 /// The valid values of index are [0,n-1] for a list of n elements,
@@ -61,7 +61,7 @@ elem_t ioopm_linked_list_remove(ioopm_list_t *list, int index, bool *success);
 /// @param index the position in the list
 /// @param success a stack variable pointer to indicate success or failure
 /// @return the value at the given position
-elem_t ioopm_linked_list_get(ioopm_list_t *list, int index, bool *success);
+elem_t ioopm_linked_list_get(ioopm_list_t *list, size_t index, bool *success);
 
 /// @brief Returns element if found in list
 /// @param list the linked list
